examples/vit-finetune/requirements.txt
flwr[simulation]>=1.0, <2.0
flwr-datasets[vision]>=0.0.2, <1.0.0
matplotlib==3.8.3